Add tab completion functionality for add command
Currently, users have to manually key in all prefixes when using add command to add a new recipient.

This results in tediousness in keying in new recipients and spends time which should be saved in this context of quickly finding a recipient for an organ donor.

Let us create a tab key press feature that sets up all the prefixes for the add command.

Now the user does not have to key in prefixes manually and can save time when adding recipients to the directory.