Add priority matching predicate
There is no way to search for persons by priority.

This means that a user is unable to perform database checks or
maintain the organ recipient priority distribution if they
wish to do so.

Let's add a predicate that can be used to filter persons by
priority.

It is done as a separate predicate so I can use it for a search
command in the future.

Implementation is similar to the blood type predicate.