Update user guide UI photo
The user guide UI photo is set to a mockup of organizer, and is out
of date.

Let's update the photo so it actually reflects what the app looks
like.

The change should not affect UserGuide.md since Ui.png is directly
replaced.

This should avoid future confusion among our users.