Allow duplicate persons
Recipients of the same name cannot be added to the address book.

Hospitals may deal with patients with identical names, and this
would result in an issue.

Let's identify users uniquely by their name and phone numbers.

Doing so allows children who might not have phone numbers to be
tagged to their parents' numbers instead.